Parse pkt-lines only, skip PACK binary section.
The major goal of this refactor is to avoid a very obvious problem that
the receivePackRegex regexp is used on all of receive-pack data,
including the pkt-lines but also binary PACK section. This can be a
problem for large git push operations that include PACK section sizes of
50 MB, 100 MB, or any arbitrarily greater size. Running a regexp over all
that binary data can be very expensive, and it's completely unnecessary
since the matches can only occur within the small pkt-lines section (on
the order of a few KB at the most).

Fixes #16.
Fixes #14.